Exercise 8 finished 9 partially finished
When you keep on finishing this Exercise, consider that
creating a list of tuples to compare the items is a very
messy solution. It is not clear how to handle the last element
and currently the last result array gets skipped.

So a better approach will be, to store the last operated value as
Maybe in a fitting data structure (Tuple with more then two elements or
RecordPattern CustomType, worst solution PairTuple of PairTuples)

Then we can easily assume, that the first element has to be part of
the result, even if its previous value is Nothing.